move operators to member, add explicit sign conversion cast

+++ Scheduler/src/job_info.cpp
@@ -33,7 +33,7 @@ bool job_info::can_run(const resource_info &resc) const noexcept(true) {
 }
 
 int job_info::fitness(const resource_info &resc) const noexcept(false) {
-	if(can_run(resc)) return resc.cores - req_resc.cores;
+	if(can_run(resc)) return static_cast<int>(resc.cores) - static_cast<int>(req_resc.cores);
 	else throw nullptr; // actual exception doesn't matter, and this is quite fast
 }
